Initial work on Azure blob store artifact repo
